+++ AppUI/Clients/Models/CardAspect.py
@@ -12,6 +12,21 @@ class CardAspect(str, Enum):
     HEROISM = 'heroism'
     VILLAINY = 'villainy'
     
+    @property
+    def rank(self) -> int:
+        return self._rank_mapping()[self]
+    
+    @staticmethod
+    def _rank_mapping() -> Dict['CardAspect', int]:
+        return {
+            CardAspect.VIGILANCE: 1,
+            CardAspect.COMMAND: 2,
+            CardAspect.AGGRESSION: 3,
+            CardAspect.CUNNING: 4,
+            CardAspect.HEROISM: 5,
+            CardAspect.VILLAINY: 6
+        }
+    
     @staticmethod
     def _emoji_mapping() -> Dict['CardAspect', str]:
         return {