+++ R4UI/BoxLayouts.py
@@ -1,14 +1,18 @@
 from typing import List
-from PyQt5.QtCore import Qt
-from PyQt5.QtWidgets import (QGridLayout, QHBoxLayout, QSpacerItem,
-                             QVBoxLayout, QWidget, QBoxLayout)
 
-class BoxLayout(QWidget):
+from PyQt5.QtCore import Qt
+from PyQt5.QtWidgets import (QBoxLayout, QGridLayout, QHBoxLayout, QSpacerItem,
+                             QVBoxLayout)
+
+from .R4UIWidget import R4UIWidget
+
+
+class BoxLayout(R4UIWidget):
     def __init__(self, 
                  layout: QBoxLayout,  
-                 widgets: List[QWidget] = []):
+                 widgets: List[R4UIWidget] = []):
         super().__init__()
-        self._widgets: List[QWidget] = []
+        self._widgets: List[R4UIWidget] = []
         self._layout = layout
         self.setLayout(self._layout)
         self.add_widgets(widgets)
@@ -21,11 +25,11 @@ class BoxLayout(QWidget):
     def set_alignment_top(self) -> 'BoxLayout':
         return self.set_alignment_for_all_widgets(Qt.AlignmentFlag.AlignTop)
     
-    def add_widgets(self, widgets: List[QWidget]):
+    def add_widgets(self, widgets: List[R4UIWidget]):
         for w in widgets:
             self.add_widget(w)
             
-    def add_widget(self, widget: QWidget):
+    def add_widget(self, widget: R4UIWidget):
         self._widgets.append(widget)
         self._layout.addWidget(widget)
     
@@ -33,7 +37,7 @@ class BoxLayout(QWidget):
         self._layout.addSpacerItem(spacer_item)
         return self
         
-    def set_to_layout(self, layout: QWidget) -> 'BoxLayout':
+    def set_layout_to_widget(self, layout: R4UIWidget) -> 'BoxLayout':
         layout.setLayout(self.layout())
         return self
     
@@ -58,11 +62,11 @@ class BoxLayout(QWidget):
                     widget = None
         self._widgets = []
                     
-    def replace_all_widgets(self, widgets: List[QWidget]):
+    def replace_all_widgets(self, widgets: List[R4UIWidget]):
         self._clear_widgets()
         self.add_widgets(widgets)
     
-    def insert_widget(self, index: int, widget: QWidget):
+    def insert_widget(self, index: int, widget: R4UIWidget):
         self._layout.insertWidget(index, widget)
      
     def remove_widget_at_index(self, index_to_remove: int):
@@ -97,25 +101,25 @@ class BoxLayout(QWidget):
         self._layout.insertItem(index_2, item_1)
 
 class HorizontalBoxLayout(BoxLayout):
-    def __init__(self, widgets: List[QWidget] = []):
+    def __init__(self, widgets: List[R4UIWidget] = []):
         super().__init__(QHBoxLayout(), widgets)
         pass
     
             
 class VerticalBoxLayout(BoxLayout):
-    def __init__(self, widgets: List[QWidget] = []):
+    def __init__(self, widgets: List[R4UIWidget] = []):
         super().__init__(QVBoxLayout(), widgets)
         pass
 
-class GridLayout(QWidget):
-    def __init__(self, widgets: List[tuple[QWidget, tuple[int, int]]] = []):
+class GridLayout(R4UIWidget):
+    def __init__(self, widgets: List[tuple[R4UIWidget, tuple[int, int]]] = []):
         super().__init__()
-        self._widgets: List[QWidget] = []
+        self._widgets: List[R4UIWidget] = []
         self._layout = QGridLayout()
         self.setLayout(self._layout)
         self.add_widgets(widgets)
         
-    def add_widgets(self, widgets: List[tuple[QWidget, tuple[int, int]]]):
+    def add_widgets(self, widgets: List[tuple[R4UIWidget, tuple[int, int]]]):
         for w, p in widgets:
             self._widgets.append(w)
             self._layout.addWidget(w, p[0], p[1])
@@ -130,11 +134,11 @@ class GridLayout(QWidget):
                     widget = None
         self._widgets = []
                     
-    def replace_all_widgets(self, widgets: List[tuple[QWidget, tuple[int, int]]]):
+    def replace_all_widgets(self, widgets: List[tuple[R4UIWidget, tuple[int, int]]]):
         self._clear_widgets()
         self.add_widgets(widgets)
         
-    def set_to_layout(self, layout: QWidget) -> 'GridLayout':
+    def set_layout_to_widget(self, layout: R4UIWidget) -> 'GridLayout':
         layout.setLayout(self.layout())
         return self